+++ src/Nacara/js/nodemon-watch.js
@@ -1,0 +1,3 @@
+import { start } from '../dist/Main.js';
+
+start();